Fix issue with targeting folder with no files
An issue was coming up in generating a checksum file for a folder with
no files in it, but files in subdirectories of the folder. This had to
do with the confusing syntax required with using the -Exclude parameter
with Get-ChildItem.

Now I know that it doesn't need to be used when you're recursing, and it
does need to be used when you aren't.